pathfinding: Widen the feature mask to 16 bits
Every bit of the byte is spoken for. Widen game.pathfindingBackend and
PathfindingFeature to 16 bits, along with the savegame and lobby
carries and the command line range check.

This changes the wire format: the option is written with NETuint16_t
now rather than NETuint8_t, so clients on the old format will not read
it correctly.

No behavior change.

+++ src/game.cpp
@@ -4802,7 +4802,7 @@ static bool loadMainFile(const std::string &fileName)
 		game.blindMode = static_cast<BLIND_MODE>(save.value("blindMode").toInt());
 	}
 	// default to the legacy A* planner for saves written before this setting existed
-	game.pathfindingBackend = static_cast<uint8_t>(save.value("pathfindingBackend", 0).toUInt());
+	game.pathfindingBackend = static_cast<uint16_t>(save.value("pathfindingBackend", 0).toUInt());
 	if (save.contains("multiplayer"))
 	{
 		bMultiPlayer = save.value("multiplayer").toBool();
